[bugfix] Add sessionOdb.upgradeOdb method
